+++ src/bin/llhd-opt.rs
@@ -110,6 +110,10 @@ fn main_inner() -> Result<(), String> {
     let t6 = time::precise_time_ns();
     llhd::pass::DeadCodeElim::run_on_module(&ctx, &mut module);
     let t7 = time::precise_time_ns();
+    llhd::pass::ControlFlowSimplification::run_on_module(&ctx, &mut module);
+    let t8 = time::precise_time_ns();
+    llhd::pass::DeadCodeElim::run_on_module(&ctx, &mut module);
+    let t9 = time::precise_time_ns();
 
     // Verify modified module.
     let mut verifier = Verifier::new();
@@ -117,7 +121,7 @@ fn main_inner() -> Result<(), String> {
     verifier
         .finish()
         .map_err(|errs| format!("Verification failed after optimization:\n{}", errs))?;
-    let t8 = time::precise_time_ns();
+    let t10 = time::precise_time_ns();
 
     // Write the output.
     if let Some(path) = matches.value_of("output") {
@@ -127,7 +131,7 @@ fn main_inner() -> Result<(), String> {
     } else {
         llhd::assembly::write_module(std::io::stdout().lock(), &module);
     }
-    let t9 = time::precise_time_ns();
+    let t11 = time::precise_time_ns();
 
     // Print execution time statistics if requested by the user.
     if matches.is_present("time-passes") {
@@ -139,9 +143,11 @@ fn main_inner() -> Result<(), String> {
         eprintln!("  TCM:     {:8.3} ms", (t5 - t4) as f64 * 1.0e-6);
         eprintln!("  LICM:    {:8.3} ms", (t6 - t5) as f64 * 1.0e-6);
         eprintln!("  DCE:     {:8.3} ms", (t7 - t6) as f64 * 1.0e-6);
-        eprintln!("  Verify:  {:8.3} ms", (t8 - t7) as f64 * 1.0e-6);
-        eprintln!("  Output:  {:8.3} ms", (t9 - t8) as f64 * 1.0e-6);
-        eprintln!("  Total:   {:8.3} ms", (t9 - t0) as f64 * 1.0e-6);
+        eprintln!("  CFS:     {:8.3} ms", (t8 - t7) as f64 * 1.0e-6);
+        eprintln!("  DCE:     {:8.3} ms", (t9 - t8) as f64 * 1.0e-6);
+        eprintln!("  Verify:  {:8.3} ms", (t10 - t9) as f64 * 1.0e-6);
+        eprintln!("  Output:  {:8.3} ms", (t11 - t10) as f64 * 1.0e-6);
+        eprintln!("  Total:   {:8.3} ms", (t11 - t0) as f64 * 1.0e-6);
     }
 
     // Dump some threading statistics.