+++ src/pass/mod.rs
@@ -6,12 +6,14 @@
 //! intermediate representation.
 
 pub mod cf;
+pub mod cfs;
 pub mod dce;
 pub mod gcse;
 pub mod licm;
 pub mod tcm;
 
 pub use cf::ConstFolding;
+pub use cfs::ControlFlowSimplification;
 pub use dce::DeadCodeElim;
 pub use gcse::GlobalCommonSubexprElim;
 pub use licm::LoopIndepCodeMotion;