+++ CHANGELOG.md
@@ -25,6 +25,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Add Temporal Code Motion pass.
 - Add `phi` instruction.
 - Add Loop Independent Code Motion Pass.
+- Add Control Flow Simplification Pass.
 
 ### Changed
 - Add missing `dyn` keywords.